Update to support oauth2
Updated to support oauth2. Unfortunately, I also switched it from the
weird indention to 2 space indention, so it's difficult to figure out
what changed in this commit.

Basically, the code that requested the oauth token was replaced with
http requests for the oauth2 token. The oauth tokens in the config file
were also replaced with the refresh token. This token changes every time
the script is run and may cause issues.